requirements.txt
alembic==1.8.0
appdirs==1.4.3
APScheduler==3.9.1
asgiref==3.5.2
autopep8==1.6.0
Babel==2.10.3
bleach==3.2.0
blinker==1.4
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.0
click==7.1.2
colorama==0.4.5
coverage==5.3
cryptocompare==0.7.3
cryptography==37.0.4
cycler==0.10.0
distlib==0.3.0
Django==4.1.1
djangorestframework==3.13.1
dnspython==1.16.0
dominate==2.6.0
ecdsa==0.18.0
email-validator==1.1.1
filelock==3.0.12
Flask==1.1.4
Flask-APScheduler==1.12.4
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.1.0
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-Phrase==1.0.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-SSLify==0.1.5
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
httpie==2.2.0
idna==2.10
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.3
jwt==1.3.1
kiwisolver==1.3.1
login==0.0.6
Mako==1.2.1
Markdown==3.2.2
MarkupSafe==1.1.1
numpy==1.23.1
packaging==20.4
Pillow==8.0.1
psycopg2==2.9.3
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.15.0
Pygments==2.7.1
PyGreSQL==5.2.1
pyparsing==2.4.7
python-bitcoinlib==0.11.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-editor==1.0.4
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.24.0
six==1.16.0
SQLAlchemy==1.4.39
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.25.10
virtualenv==20.0.10
visitor==0.1.3
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==3.0.1
zipp==3.1.0
